+++ libraries/resource_providers/api200/logical_interconnect_provider.rb
@@ -51,7 +51,7 @@ module OneviewCookbook
       # Handle the most types of updates
       def update_handler(action, key = nil)
         temp = key ? { key.to_s => Marshal.load(Marshal.dump(@item[key])) } : Marshal.load(Marshal.dump(@item.data))
-        raise "Resource not found: Action '#{action}' cannot be performed since #{@resource_name} '#{@name}' was not found." unless @item.retrieve!
+        @item.retrieve! || raise("#{@resource_name} '#{@name}' not found!")
         return Chef::Log.info("#{@resource_name} '#{@name}' is up to date") if @item.like?(temp)
         diff = get_diff(@item, temp)
         diff.insert(0, '. Diff:') unless diff.to_s.empty?
@@ -66,7 +66,7 @@ module OneviewCookbook
       end
 
       def firmware_handler(action)
-        @item.retrieve!
+        @item.retrieve! || raise("#{@resource_name} '#{@name}' not found!")
         current_firmware = @item.get_firmware
         @context.firmware_data['command'] = action
         dif_values = @context.firmware_data.select { |k, v| current_firmware[k] != v }
@@ -111,7 +111,7 @@ module OneviewCookbook
       end
 
       def update_internal_networks
-        @item.retrieve!
+        @item.retrieve! || raise("#{@resource_name} '#{@name}' not found!")
         @context.internal_networks.collect! { |n| load_resource(:EthernetNetwork, n) }
         if @item['internalNetworkUris'].sort == @context.internal_networks.collect { |x| x['uri'] }.sort
           Chef::Log.info("Internal networks for #{@resource_name} #{@name} are up to date")
@@ -170,7 +170,7 @@ module OneviewCookbook
       end
 
       def update_from_group
-        @item.retrieve!
+        @item.retrieve! || raise("#{@resource_name} '#{@name}' not found!")
         # Nothing to verify
         @context.converge_by "Update #{@resource_name} '#{@name}' from group" do
           @item.compliance
@@ -178,7 +178,7 @@ module OneviewCookbook
       end
 
       def reapply_configuration
-        @item.retrieve!
+        @item.retrieve! || raise("#{@resource_name} '#{@name}' not found!")
         # Nothing to verify
         @context.converge_by "Reapply configuration in #{@resource_name} '#{@name}'" do
           @item.configuration