+++ spec/unit/resources/logical_interconnect/update_firmware_spec.rb
@@ -21,4 +21,9 @@ describe 'oneview_test::logical_interconnect_update_firmware' do
       .with('Update', @firmware_driver, anything).and_return(true)
     expect(real_chef_run).to update_oneview_logical_interconnect_firmware('LogicalInterconnect-update_firmware')
   end
+
+  it 'fails if the resource is not found' do
+    expect_any_instance_of(OneviewSDK::LogicalInterconnect).to receive(:retrieve!).and_return(false)
+    expect { real_chef_run }.to raise_error(RuntimeError, /not found/)
+  end
 end